requirements.txt
pykaldi
tqdm
pandas
matplotlib
seaborn
librosa
sox
pysubs2
flask